rework all "duration" values into the actual `Duration` type and parse them dynamically (#1739)
* rework all "duration" values into the actual `Duration` type and parse them dynamically

* fix broken OTP exp check

* update docs

* update book pages with duration value types

* revert to `hiqlite` from `main` until release

+++ src/api/src/fed_cm.rs
@@ -435,7 +435,7 @@ async fn login_status_from_req(req: &HttpRequest) -> (FedCMLoginStatus, String) 
             };
 
             if session.is_valid(
-                RauthyConfig::get().vars.fedcm.session_timeout,
+                RauthyConfig::get().vars.fedcm.session_timeout as i64,
                 real_ip_from_req(req).ok(),
                 req.path(),
             ) {
